+++ lib/components/welcome_button.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+
+class WelcomeButton extends StatelessWidget {
+  final String buttonText;
+  final Widget onTap;
+  final Color color;
+  final Color textColor;
+
+  const WelcomeButton({
+    required this.buttonText,
+    required this.onTap,
+    required this.color,
+    required this.textColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      hoverColor: Colors.transparent,
+      borderRadius: BorderRadius.circular(50),
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (e) => onTap,
+          ),
+        );
+      },
+      child: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 30),
+        child: Container(
+          alignment: Alignment.center,
+          height: 50,
+          decoration: BoxDecoration(
+            color: color,
+            borderRadius: BorderRadius.circular(50),
+            boxShadow: [
+              BoxShadow(
+                color: Colors.black.withOpacity(0.2),
+                spreadRadius: 2,
+                blurRadius: 5,),]
+            ),
+          child: Text(
+            buttonText,
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              fontSize: 20.0,
+              fontWeight: FontWeight.bold,
+              color: textColor,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
