--- lib/holder/login/login_page.dart
@@ -1,296 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../components/Logo.dart';
-import '../../components/custom_form.dart';
-import '../../components/login_button.dart';
-import '../../size.dart';
-import '../../theme.dart';
-import '../join/join_page.dart';
-
-class LoginPage extends StatelessWidget {
-  const LoginPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: ListView(
-        children: [
-          SizedBox(height: xlargeGap),
-          Logo(title: "loging"),
-          CustomForm(),
-          Row(
-            children: [
-              const Expanded(
-                child: WelcomeButton(
-                  buttonText: '로그인',
-                  onTap: SignInScreen(),
-                  color: Colors.transparent,
-                  textColor: Colors.white,
-                ),
-              ),
-              Expanded(
-                child: WelcomeButton(
-                  buttonText: '회원가입',
-                  onTap: const joinPage(),
-                  color: Colors.white,
-                  textColor: lightColorScheme.primary,
-                ),
-              ),
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class SignInScreen extends StatefulWidget {
-  const SignInScreen({super.key});
-
-  @override
-  State<SignInScreen> createState() => _SignInScreenState();
-}
-
-class _SignInScreenState extends State<SignInScreen> {
-  final _formSignInKey = GlobalKey<FormState>();
-  bool rememberPassword = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        const Expanded(
-          flex: 1,
-          child: SizedBox(
-            height: 10,
-          ),
-        ),
-        Expanded(
-          flex: 7,
-          child: Container(
-            padding: const EdgeInsets.fromLTRB(25.0, 50.0, 25.0, 20.0),
-            decoration: const BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.only(
-                topLeft: Radius.circular(40.0),
-                topRight: Radius.circular(40.0),
-              ),
-            ),
-            child: SingleChildScrollView(
-              child: Form(
-                key: _formSignInKey,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Text(
-                      'Welcome back',
-                      style: TextStyle(
-                        fontSize: 30.0,
-                        fontWeight: FontWeight.w900,
-                        color: lightColorScheme.primary,
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 40.0,
-                    ),
-                    TextFormField(
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Please enter Email';
-                        }
-                        return null;
-                      },
-                      decoration: InputDecoration(
-                        label: const Text('Email'),
-                        hintText: 'Enter Email',
-                        hintStyle: const TextStyle(
-                          color: Colors.black26,
-                        ),
-                        border: OutlineInputBorder(
-                          borderSide: const BorderSide(
-                            color: Colors.black12, // Default border color
-                          ),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                        enabledBorder: OutlineInputBorder(
-                          borderSide: const BorderSide(
-                            color: Colors.black12, // Default border color
-                          ),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    TextFormField(
-                      obscureText: true,
-                      obscuringCharacter: '*',
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Please enter Password';
-                        }
-                        return null;
-                      },
-                      decoration: InputDecoration(
-                        label: const Text('Password'),
-                        hintText: 'Enter Password',
-                        hintStyle: const TextStyle(
-                          color: Colors.black26,
-                        ),
-                        border: OutlineInputBorder(
-                          borderSide: const BorderSide(
-                            color: Colors.black12, // Default border color
-                          ),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                        enabledBorder: OutlineInputBorder(
-                          borderSide: const BorderSide(
-                            color: Colors.black12, // Default border color
-                          ),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Row(
-                          children: [
-                            Checkbox(
-                              value: rememberPassword,
-                              onChanged: (bool? value) {
-                                setState(() {
-                                  rememberPassword = value!;
-                                });
-                              },
-                              activeColor: lightColorScheme.primary,
-                            ),
-                            const Text(
-                              'Remember me',
-                              style: TextStyle(
-                                color: Colors.black45,
-                              ),
-                            ),
-                          ],
-                        ),
-                        GestureDetector(
-                          child: Text(
-                            'Forget password?',
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              color: lightColorScheme.primary,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    SizedBox(
-                      width: double.infinity,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          if (_formSignInKey.currentState!.validate() &&
-                              rememberPassword) {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(
-                                content: Text('Processing Data'),
-                              ),
-                            );
-                          } else if (!rememberPassword) {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(
-                                  content: Text(
-                                      'Please agree to the processing of personal data')),
-                            );
-                          }
-                        },
-                        child: const Text('Sign up'),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Expanded(
-                          child: Divider(
-                            thickness: 0.7,
-                            color: Colors.grey.withOpacity(0.5),
-                          ),
-                        ),
-                        const Padding(
-                          padding: EdgeInsets.symmetric(
-                            vertical: 0,
-                            horizontal: 10,
-                          ),
-                          child: Text(
-                            'Sign up with',
-                            style: TextStyle(
-                              color: Colors.black45,
-                            ),
-                          ),
-                        ),
-                        Expanded(
-                          child: Divider(
-                            thickness: 0.7,
-                            color: Colors.grey.withOpacity(0.5),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    const SizedBox(
-                      height: 25.0,
-                    ),
-                    // don't have an account
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        const Text(
-                          'Don\'t have an account? ',
-                          style: TextStyle(
-                            color: Colors.black45,
-                          ),
-                        ),
-                        GestureDetector(
-                          onTap: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (e) => const joinPage(),
-                              ),
-                            );
-                          },
-                          child: Text(
-                            'Sign up',
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                              color: lightColorScheme.primary,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(
-                      height: 20.0,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
