+++ lib/holder/main_holder.dart
@@ -1,4 +1,4 @@
-import 'package:final_project_team02/holder/login/login_page.dart';
+import 'package:final_project_team02/holder/welcome/welcome_page.dart';
 import 'package:flutter/material.dart';
 
 class MainHolder extends StatelessWidget {
@@ -8,7 +8,7 @@ class MainHolder extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      body: LoginPage(),
+      body: WelcomePage(),
     );
   }
 }
